adding pricing service and decorator

+++ parking-lot/src/main/java/com/lld/strategies/price/BikePriceStrategy.java
@@ -2,7 +2,7 @@ package com.lld.strategies.price;
 
 public class BikePriceStrategy implements PriceCalculationStrategy {
     @Override
-    public float calculatePrice(int minutes) {
-        return (float) 50 * minutes / 60;
+    public double calculatePrice(long minutes) {
+        return (double) 50 * minutes / 60;
     }
 }
