+++ parking-lot/src/main/java/com/lld/strategies/price/CarPriceStrategy.java
@@ -3,7 +3,7 @@ package com.lld.strategies.price;
 public class CarPriceStrategy implements PriceCalculationStrategy {
 
     @Override
-    public float calculatePrice(int minutes) {
-        return (float) 100 * minutes / 60;
+    public double calculatePrice(long minutes) {
+        return (double) 100 * minutes / 60;
     }
 }
